requirement.txt
Django==3.0.6
djangorestframework
Pillow
requests
